# Faster int_to_big_endian implementation