--- a/eth_utils/encoding.py
+++ b/eth_utils/encoding.py
@@ -1,10 +1,6 @@
-import math
-
-
 def int_to_big_endian(value):
-    byte_length = max(math.ceil(value.bit_length() / 8), 1)
-    return value.to_bytes(byte_length, byteorder='big')
+    return value.to_bytes((value.bit_length() + 7) // 8 or 1, 'big')
 
 
 def big_endian_to_int(value):
-    return int.from_bytes(value, byteorder='big')
+    return int.from_bytes(value, 'big')