@davesque 2018-04-18
@pipermerriam Here's something interesting:

```python
from math import ceil
import collections
import math
import random
import time

def profile(to_wrap, runs=100000):
    start = time.time()

    for i in range(1, runs + 1):
        to_wrap(i)

    duration = time.time() - start

    return duration

def int_to_big_endian(value):
    byte_length = max(math.ceil(value.bit_length() / 8), 1)
    return value.to_bytes(byte_length, byteorder='big')

def or_int_to_big_endian(value):
    byte_length = math.ceil(value.bit_length() / 8) or 1
    return value.to_bytes(byte_length, byteorder='big')

def new_int_to_big_endian(value):
    return value.to_bytes(math.ceil(value.bit_length() / 8) or 1, byteorder='big')

def new_new_int_to_big_endian(value):
    return value.to_bytes(ceil(value.bit_length() / 8) or 1, byteorder='big')

def run():
    funcs = [
        int_to_big_endian,
        or_int_to_big_endian,
        new_int_to_big_endian,
        new_new_int_to_big_endian,
    ]
    results = collections.defaultdict(list)
    runs = 100000

    print('Performing test runs...')

    for i in range(1, 16):
        random.shuffle(funcs)
        for f in funcs:
            results[f].append(profile(f, runs=runs))

    for func, result_list in results.items():
        avg = sum(result_list) / len(result_list)

        print(f'\n~~~~~~~~~~~~ {func.__name__} ~~~~~~~~~~~~')
        print(f'{avg} s to run on average')
```
Results from a few test runs:
```python
In [2]: run()
Performing test runs...

~~~~~~~~~~~~ new_int_to_big_endian ~~~~~~~~~~~~
0.07746715545654297 s to run on average

~~~~~~~~~~~~ int_to_big_endian ~~~~~~~~~~~~
0.09578965504964193 s to run on average

~~~~~~~~~~~~ new_new_int_to_big_endian ~~~~~~~~~~~~
0.0716497262318929 s to run on average

~~~~~~~~~~~~ or_int_to_big_endian ~~~~~~~~~~~~
0.07660263379414876 s to run on average

In [3]: run()
Performing test runs...

~~~~~~~~~~~~ new_int_to_big_endian ~~~~~~~~~~~~
0.07638045946756998 s to run on average

~~~~~~~~~~~~ or_int_to_big_endian ~~~~~~~~~~~~
0.07697428067525228 s to run on average

~~~~~~~~~~~~ int_to_big_endian ~~~~~~~~~~~~
0.09517072041829427 s to run on average

~~~~~~~~~~~~ new_new_int_to_big_endian ~~~~~~~~~~~~
0.07155067125956217 s to run on average

In [4]: run()
Performing test runs...

~~~~~~~~~~~~ int_to_big_endian ~~~~~~~~~~~~
0.09540775616963705 s to run on average

~~~~~~~~~~~~ or_int_to_big_endian ~~~~~~~~~~~~
0.0766679286956787 s to run on average

~~~~~~~~~~~~ new_int_to_big_endian ~~~~~~~~~~~~
0.07332390149434408 s to run on average

~~~~~~~~~~~~ new_new_int_to_big_endian ~~~~~~~~~~~~
0.07132296562194824 s to run on average
```
I randomized the order of test runs just to lessen the likelihood that some kind of memory caching was affecting the results.  In general, it seems `new_new_int_to_big_endian` is faster.  I think the explanation might be that it avoids the implicit dict lookup that comes with dot access to get `ceil` off the `math` module object.