@davesque 2018-04-18
And here's something even more interesting.  I added another version of the `new_new_int_to_big_endian` function called `new_new_new_int_to_big_endian` that passes the byteorder argument as a positional argument instead of a keyword argument.  Here are the results for that:
```python
In [2]: run()
Performing test runs...

~~~~~~~~~~~~ new_int_to_big_endian ~~~~~~~~~~~~
0.07343041102091472 s to run on average

~~~~~~~~~~~~ or_int_to_big_endian ~~~~~~~~~~~~
0.07479095458984375 s to run on average

~~~~~~~~~~~~ int_to_big_endian ~~~~~~~~~~~~
0.09101295471191406 s to run on average

~~~~~~~~~~~~ new_new_int_to_big_endian ~~~~~~~~~~~~
0.07060823440551758 s to run on average

~~~~~~~~~~~~ new_new_new_int_to_big_endian ~~~~~~~~~~~~
0.0439151922861735 s to run on average

In [3]: run()
Performing test runs...

~~~~~~~~~~~~ new_int_to_big_endian ~~~~~~~~~~~~
0.0733139673868815 s to run on average

~~~~~~~~~~~~ or_int_to_big_endian ~~~~~~~~~~~~
0.07647164662679036 s to run on average

~~~~~~~~~~~~ new_new_new_int_to_big_endian ~~~~~~~~~~~~
0.043297338485717776 s to run on average

~~~~~~~~~~~~ int_to_big_endian ~~~~~~~~~~~~
0.09046815236409506 s to run on average

~~~~~~~~~~~~ new_new_int_to_big_endian ~~~~~~~~~~~~
0.07095227241516114 s to run on average

In [4]: run()
Performing test runs...

~~~~~~~~~~~~ int_to_big_endian ~~~~~~~~~~~~
0.09290258089701335 s to run on average

~~~~~~~~~~~~ new_new_new_int_to_big_endian ~~~~~~~~~~~~
0.04409352938334147 s to run on average

~~~~~~~~~~~~ or_int_to_big_endian ~~~~~~~~~~~~
0.07707031567891438 s to run on average

~~~~~~~~~~~~ new_new_int_to_big_endian ~~~~~~~~~~~~
0.07176694869995118 s to run on average

~~~~~~~~~~~~ new_int_to_big_endian ~~~~~~~~~~~~
0.07464202245076497 s to run on average
```