@pipermerriam 2018-04-18
holy crap, the positional argument thing dropped like anothe 25% off the total time.  Latest version of this is more than twice as fast (each run is 40% of the original execution time).

Good finds @davesque