@pipermerriam 2018-04-18
I was able to drop the call to `math.ceil` with `(value.bit_length() + 7) // 8`